pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "crazy-rl"
version = "0.0.1"
description = "Environments and learning code for doing (MOMA)RL with CrazyFlies."
authors = ["Florian Felten <florian.felten@uni.lu>"]
license = "MIT"
readme = "README.md"
keywords = ["Reinforcement Learning", "Multi-Objective", "Multi-Agent", "RL", "AI", "gymnasium", "pettingzoo"]
classifiers = [
    "Development Status :: 4 - Beta",  # change to `5 - Production/Stable` when ready
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    'Intended Audience :: Science/Research',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
]

[tool.poetry.dependencies]
python = "^3.9, <3.12"
gymnasium = "^0.28"
pettingzoo = "^1.23"
numpy = "*"
torch = "^1.13, <2.0.0"
pygame = "^2.5"
pyopengl = "==3.1.6"
cflib = "^0.1"
pytest = "^7.1"
jax = {version="^0.4"}
jaxlib = "^0.4"
flax = "^0.6"
optax = "^0.1"
chex = "^0.1"
orbax-checkpoint = "^0.3"
matplotlib = "^3.7"
distrax = {git = "https://github.com/deepmind/distrax.git", rev = "93c54a8b15c07e57a58b95b519e5510ab99220c9"}
jax_dataclasses = "^1.5"
pymoo = "^0.6"
mplcursors = "^0.5.2"
supersuit = "^3.9.0"
tqdm = "^4.66.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20"

[tool.poetry.extras]
pytest = ["pytest"]

# Linting, testing, ... ########################################################
[tool.black]
safe = true
line-length = 127
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
atomic = true
profile = "black"
src_paths = ["crazy_rl"]
extra_standard_library = ["typing_extensions"]
indent = 4
lines_after_imports = 2
multi_line_output = 3
